+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 - Massimport deeplink support [@mrissaoussama](https://github.com/mrissaoussama) [#409](https://github.com/tsundoku-otaku/tsundoku/pull/409)
 - Add option to clear only error-free completed massimport batches [@mrissaoussama](https://github.com/mrissaoussama) [#402](https://github.com/tsundoku-otaku/tsundoku/pull/402)
 - Add multiline LLM prompt editor, and reset option to all LLM options [@mrissaoussama](https://github.com/mrissaoussama) [#411](https://github.com/tsundoku-otaku/tsundoku/pull/411)
+- Add more options for tap zones, adjustable heights ; fix autosplit [@mrissaoussama](https://github.com/mrissaoussama) [#413](https://github.com/tsundoku-otaku/tsundoku/pull/413)
 
 
 ### Fixed